Remove all references to the ConfirmHelper and the inputs from the new enketo directive
We could theoretically replace the user input array by a single value but we
retain it as an array with a single, hardcoded value. We do remove the
iteration of the array and access it directly using the single, hard-coded
value. This allows us to support incremental changes, and defer the more
complex refactoring until later.

Similarly, we add early returns to all the verification code to convert it to a
NOP and ensure that it is not used. We should remove it before finalizing this
PR since we don't want to have copy-pasted and unused code lurking around in
the codebase. We can restore it from the multi-label-ui codebase when we do
start supporting guesses for the surveys as well.

```
$ grep -r ConfirmHelper www/js/survey/enketo/
Binary file www/js/survey/enketo//.enketo-trip-button.js.swp matches
```