Remove all references to the ConfirmHelper from the diary services
Instead, define `MANUAL_KEYS` in each directive and iterate over them.
Also, it turned out that this was not working earlier
https://github.com/e-mission/e-mission-docs/issues/727#issuecomment-1121933037
not because of any timing issues, but because we were reading the wrong key

TODO: see if we can combine `extractResult` and `processManualInputs` since the
timing was not the issue after all. However, the timing *could* be an issue
since it is async, and we don't want to rely on timing effects, so we may also
want to leave it unchanged.
`