Revert unncessary changes introduced by the enketo code
The previous implementation treated the enketo code as another type of survey.
However the enketo survey is different than the multi-label UI.
We have now separated out the functionality into separate directives, and removed all
references to the ConfirmHelper in the diary code

The eneketo survey is a single array entry with key "SURVEY", and is stored in the directive.
In the diary, deployers can choose between the multi-ui and enketo implementations
https://github.com/e-mission/e-mission-docs/issues/727#issuecomment-1120254533