Refactor the answer matching code for the survey
Since we already match the user input for the trip during
populateInputsAndInferences, we don't need to re-pull all the entries and
re-match during the survey launch.

This is likely to be particularly bad if we collect data over a long period of
time. Reading all the user inputs for a year every time we want to launch the
survey is quickly going to get old.

This also means that we need to store the entire survey object instead of only the label
when we match entries.
https://github.com/e-mission/e-mission-docs/issues/727#issuecomment-1122050160