once the intro is done, remove the intro view
If the user "logs out" and goes back to the intro screen, it is re-created.
This ensures that we don't have two form fields sitting around right after we
have finished the onboarding, and that opening the popup will not add a new
form to the wrong form field

This fixes https://github.com/e-mission/e-mission-docs/issues/727#issuecomment-1125168736
https://github.com/e-mission/e-mission-docs/issues/727#issuecomment-1125218538
https://github.com/e-mission/e-mission-docs/issues/727#issuecomment-1125443770
https://github.com/e-mission/e-mission-docs/issues/727#issuecomment-1125443770

We tried to remove only the form directive but in that case, the form is not
re-created if we go back to the intro screen.
https://github.com/e-mission/e-mission-docs/issues/727#issuecomment-1125457696

Deleting the entire view does work
https://github.com/e-mission/e-mission-docs/issues/727#issuecomment-1125473618