Remove "performance hack" for the user matching
Where if the user input list has a single entry, we can return it.
If we have one user input and multiple trips, we can't just return that one
input. That is the input for one trip, not the others!
https://github.com/e-mission/e-mission-docs/issues/727#issuecomment-1139200320

wrt performance, we already handled the major performance issue by only having
to manually match unprocessed entries after the end of the pipeline

Testing done:
- loaded a day with a single user input
- before this change, all trips were flagged with that input
- after this change, only the first was