Fix the promises code to retrieve unprocessed labels
Before this, we used to return the Promises.all and fill in the map in its
then.

As part of the refactoring for the enketo trip labels,
https://github.com/e-mission/e-mission-phone/commit/aec00514af167f1d99465110d03ba14596c392a5

we changed it to return an array of the result and the Promise.all

This meant that the infinite scroll code, which was expecting a result map, got
a promise array instead, and so the mapping of user inputs in the infinite
scroll list = label list failed consistently.

Fixing this by re-returning `Promise.all` and the result and mapping from the `then`

Note that the labels from the diary were not visible on the label screen even for the classic multilabel UI
https://github.com/e-mission/e-mission-docs/issues/727#issuecomment-1139200320

Original code:

```

            return Promise.all(manualPromises).then((manualResults) => {
                const manualConfirmResults = {};
                manualResults.forEach(function(mr, index) {
                  manualConfirmResults[ConfirmHelper.INPUTS[index]] = mr;
                });
                return [result, manualConfirmResults];
            });
```

Broken code:

```
            return [result, Promise.all(manualPromises).then((manualResults) =>
                manualInputFactory.processManualInputs(manualResults, manualConfirmResults))];
```

Fixed code:

```
            return Promise.all(manualPromises).then((manualResults) => {
                manualInputFactory.processManualInputs(manualResults, manualConfirmResults);
                return [result, manualConfirmResults];
            });
```