Ensure that we load the survey in the phone language
Notably, when the phone language is Spanish, we want to load the spanish
version of the survey. This is apparently functionality that is in the most
recent version of enketo, but not in our older version.

We can apply the patch that introduced the change
https://github.com/enketo/enketo-core/commit/c31dccc811954b50bfd64be033935b8885a584f6
but we can't do it directly since we only have the bundle included as a
manually imported library in our code.

So we apply the patch by copy-pasting. This works.

Difference between versions: https://github.com/e-mission/e-mission-docs/issues/732#issuecomment-1145562589
Patch fix: https://github.com/e-mission/e-mission-docs/issues/732#issuecomment-1145567999