Load the config when the user clicks the URL + use it in the intro
This is the first part of a much more complex change.

For now, we ensure that we cache the config in both cases where the user clicks on the URL:
- previously stored config
- downloaded config

Then we generate a notification that the config has changed

Listen to this change in the intro and configure (one of) the templates accordingly

Still need to configure the other templates and the other locations where the
config is used but wanted to get this checked in first.

First step towards:
https://github.com/e-mission/e-mission-docs/issues/739#issuecomment-1153286769