Finish templatizing the other intro files as well
Replace hardcoded values throughout the consent-text with templatized values
This involved straight substitutions using `{{}}` and `<ng-if` sections
depending on whether the selected deployment is a program or a study

Also improved the formatting of the consent screen by breaking up the "wall of
text" into multiple paragraphs.

Also added emojis to demarcate lines since the current CSS doesn't seem to work
properly for `<li` items.

This is some more along the lines of:
https://github.com/e-mission/e-mission-docs/issues/739#issuecomment-1153286769

Current screenshots:
https://github.com/e-mission/e-mission-docs/issues/739#issuecomment-1153349709