Change the metric vs. imperial units to be configurable dynamically
This is probably the most requested feature ever.

We might actually want to move this to a user-defined config later,
but we still need to define a default. And since nobody likes to change
configurations, we probably need to put the default somewhere.

Best option, of course, would be to make the default locale specific
https://github.com/e-mission/e-mission-docs/issues/739#issuecomment-1153396907