Ensure that we check for remote push type only if the push format is valid
Not quite sure how this ever worked - I guess we didn't test it with invalid
format pushes.

Add braces to ensure that the remote notify types are only checked if the format is valid.
Without this, they are checked every time, which may crash if the push is of an invalid format

Encountered this because we apparently got a push of format `{'foreground': true, ...}`
Not sure who sent it, but we got a backtrace because of it