Make sure to handle studies as well
By setting the trip INPUT and inputDetails when to the full values when it is a
study

Now that we use trip-level values as opposed to controller-level values
(cd083e6c03770424c79d4230f697541f1442b6c4), we need to initialize in all cases,
not only the program.

Otherwise, we run into an issue where the trip-level inputs are undefined.