Display trip specific input options on the platform version of the app
When we started supporting configurable programs, we also wanted to show the
"replaced mode" only for the program of interest aka "e-bike" for all the
programs so far.

So we made the list of inputs trip specific instead of configured for the
entire app through a config file.

https://github.com/e-mission/e-mission-phone/commit/c0876131a550943a1a832838f7f541eab0297025
https://github.com/e-mission/e-mission-phone/commit/cd083e6c03770424c79d4230f697541f1442b6c4

But this means that we need to iterate over `trip.inputDetails` instead of `userInputDetails`
With this change, the labels are now displayed properly