fix deletion of just-added entries
`dataKey` is needed for storage in UserCache. This was being retrieved from the metadata property. However, just-added entries lack metadata
So, immediately after a survey is finished, we will include `key` in the tripAddition object along with `write_ts`
After a refresh, this will disappear and we will have `metadata.key` instead