Bump up versions of the dev chain and all other plugins
- android CLI version to the latest
- android sdk packages to the latest
- cordova, ionic and bower to the latest
- all non-OpenPATH plugins to the latest versions

Some of these (`cordova-plugin-badge`/`phonegap-plugin-barcodescanner`) ran into an issue with using `compile` when they actually needed `implementation` for the new gradle.

The badge is a dependency of the local notification plugin, so I migrated from `timkellypa` to `levanax` who has some commits that have fixed it and some other issues. This fixes it for the local notification plugin but not for the `cordova-plugin-badge` which is pulled in as a dependency from elsewere

I fixed that and the barcodescanner by adding a new hook that renames `compile` to `implementation` before build.

Testing done: builds