Revert the QR code plugin to the one that we used before
We tried to migrate from `phonegap-plugin-barcodescanner` to the better
supported `cordova-plugin-barcode-qrscanner`. However, I was not able to get
the new plugin to work. The OS indicated that the app was accessing the camera,
but we were not able to see the preview, even after I added a `show` to the code.

https://github.com/e-mission/e-mission-docs/issues/838#issuecomment-1474273170
https://github.com/e-mission/e-mission-docs/issues/838#issuecomment-1474344148

There doesn't seem to be an option to file issues or to communicate with the creator.
So let's fall back to the older `phonegap-plugin-barcodescanner`
We restore the `android_change_compile_implementation` hook so that we can
change the `compile` to `implementation` and get the plugin to compile.

Bonus fix: Add the `NSCameraUsageDescription` via the config.xml so that we
don't get a permission issue while running on iOS.

Testing done:
- Builds on both android and iOS
- Scanning launches on both android and iOS
- QR codes are scanned on both android and iOS