notes list CSS patch-up
Since we will left-align the label, we must remove the horizontal justification, but we still want to keep the vertical alignment in notes-list-entry. To do this, we'll use align-items instead of place-items
(place-items is shorthand for both justify-items and align-items)

Also, no need to create a whole new class for this. Let's just select the <b> inside of `.notes-list-entry` and add a helpful comment so we know what it refers to