use callback to nextEntry instead of direct reference
nextEntry is needed for input matching in specific cases. This was being put in the trip/place object, but caused deeply-nested references between trips/places, their next entry, that entry's next entry, etc...
To remedy this, we will instead use a callback function getNextEntry() that can be called in the scenarios it is needed.