Load the previous completed demographic survey
We always intended to load the previously completed demographic survey and
wrote a bunch of code to do so. However, this code attempted to load the survey
when the app is launched. With the multi-tenant architecture, we have not yet
logged in when the app is launched, so we are not able to retrieve any data
from the server.

And after we login, we don't appear to re-read the data, so the users have to
fill in the survey Every Single Time.

We tried to handle this by loading the survey after login, but it is
complicated because after the survey is read, it is stored in the controller
scope, and it is very hard to manipulate container state externally.

While looking for a simpler fix, I realized that the reason we were reading the
survey at launch was because the custom HTML element was being created at
launch as part of the intro.html.

So I:
- put an ng-if so that the survey page would only be loaded after we had a `currToken`
- converted the save token popup into another screen to slow down the user and give the survey time to load
    - since the token popup is now a "screen", it will not show the partially loaded survey behind it
    - only after the user reads and decides on the "save token", we will go to the next screen
    - at that point, the survey should have loaded

Also, made sure that the demographic survey was parsed out per the new requirements before passing it in

Testing done:
- Install the app from scratch (no cached survey in the local usercache)
- Log in to an account that has a survey stored on the server
- The QR code is displayed on a separate screen
- The survey is loaded from the server and can be edited

This fixes
https://github.com/e-mission/e-mission-docs/issues/876