fix "profile" tab directing to label screen
For some reason, clicking on the "Profile" tab would sometimes lead to the label screen.

I found that this is handled in startprefs.js getNextState(). When the result from getPendingOnboardingState() is null, the default state to return is 'root.main.inf_scroll' (aka the label screen).
The issue only happens when attempting to navigate to the Profile tab. All other tabs work consistently. So, the fallback has been changed to 'root.main.control' (aka the profile screen).

This does not address the root cause, but it works consistently and I am not going to spend more time on it because our routing system will be completely different once we migrate away from Ionic and Angular1.