route to label screen on startup and on login
As a follow-up to commit f863264f19f61ba333c8392ad91c588c94bdff7c, we do need to make some exceptions. After signing in from 'intro', we should show the label screen -- and on every subsequent startup, we land on the label screen.
Like I said in the previous commit message, this is not the cleanest solution but it fixes the issue and we will not use this routing system for much longer anyway.