fix zero-duration places showing
After any untracked time, a zero-duration place exists. We do not want to show this in the UI.
Previously, we checked the existence of place.duration and if it was less than 60 seconds. If so, it would skip this place.

However, if duration is 0, this would also be falsy in Javascript and it wouldn't skip. Instead we'll use !isNan() to make sure it is a number before doing the `<` comparison.