fix overlapping trip/place cards on android
Because of rendering differences between Android and iOS webviews, the hardcoded values here do not necessarily match the exact rendered height (it seems that the cards are rendered as taller in the Android webview than on iOS).
Ideally, our UI would be responsive and accomodate for this rather than relying on specific pixel values. But for right now, I'm just going to leave 15px as a buffer.