fix regression with zero-duration start places
Regression was introduced in https://github.com/e-mission/e-mission-phone/pull/958
In the case of zero- or short- duration start places, because of the early return statement, the corresponding trip would never be added to displayTimelineEntries.
The fix will no longer use the early return syntax.
Also removing the placeholder "12:00 AM" and "11:59 PM" since it is not necessary to show them.