:wrench: support a generic mode_studied instead of having it be hardcoded to "e-bike"
We always wanted the mode studied to be dynamically configurable, but ran out
of time/complexity, so hardcoded to e-bike. Now that we have a program that is
evaluating scootershare, we change that to make it general.

Thanks to multiple fixes in restructuring the codebase, this was a super-simple
change - we store the mode_studied and find/replace e-bike with the mode studied

Testing done:
- Loaded a program where the mode of study was "e-bike"
    - set the mode to e-bike, saw the replaced mode button
    - set the mode away from e-bike, replaced mode button went away
- Loaded a program where the mode of study was "scootershare"
    - set the mode to scootershare, saw the replaced mode button
    - set the mode away from scootershare, replaced mode button went away