:arrow_up: ensure that we remove the local version of the OPGeofenceCfg as well
In https://github.com/e-mission/e-mission-data-collection/commit/92f41145e58c49e3145a9222a78d1ccacd16d2a7,
we enabled the OPGeofence by default.
So we lazily deleted the OPGeofenceCfg from the local storage so that we don't
clutter up the local storage unnecessarily.
https://github.com/e-mission/e-mission-data-collection/commit/92f41145e58c49e3145a9222a78d1ccacd16d2a7#diff-962320754eba07107ecd413954411f725c98fd31cddbb5defd4a542d1607e5a3R160

However, because we used the kvstore to set the value, it is still in the local
storage. So users who have upgraded the app and set the value get a popup, and
the local value is copied over to remote.
https://github.com/e-mission/e-mission-phone/blob/master/www/js/plugin/storage.js#L69

- Adding a hack to remove both local and remote storage on startup so we don't
  have this popup any more
- It should now be safe to remove the backwards compat hack from the storage code,
  but I vaguely seem to have remembered seeing this, so let's keep it in until
  we have a way to auto-report it and ensure that it doesn't actually happen for
  6 months or something. Because we now rely on local storage for *everything*