ensure that notif native calls occur in sequence
With the local notification plugin, there is some latency between the end of the function calls and the completion of the tasks. The plugin functions can take callback functions as an argument that are called when the task completes.
We'll use these and integrate it with async javascript, so that `scheduleNotifs()` now returns a Promise.
Before this change, if people changed the time quickly in succession, the "cancel" requests could overlap with the "schedule" requests, causing the incorrect number of notifications to sometimes be scheduled.