prefill addition start with latest, + refactor
When additions are entered, the latest addition end time will be used to prefill the new addition's start time.
This makes sense for the common use case, where people enter additions in order chronologically.

This commit also refactors the existing logic to only prefill the time fields that are present. Since we are no longer displaying "12:00 AM" and "11:59 PM" in the place of missing timestamps, the survey timestamps should reflect this change as well.