fix final inference being overwritten by enbs
Yellow label buttons for inferred modes were not showing on the label screen for a while now.
This is because the finalInference, in the trip object, was being overwritten in 'EnketoNotesButtonService'/'enbs' (in enketo-add-note-button). This is because the logic for enketo-add-note-button was based on the logic for enketo-trip-button.
'finalInference' is not relevant to additions, so we will just remove that line.