:recycle: Change the join page to support entering the token directly
Before this, we used to support scanning the QR code (if on laptop) or clicking a link (if on phone).

Since we now want to read the token directly, we:
- change the text to indicate that they need to be scanning an opcode, and explaining what an opcode is
- offering three separate methods to enter the opcode
    - scanning a QR code
    - clicking on a link (which does not require a button)
    - copy-pasting the opcode

The scan code is unchanged, since it just needs to pass the URL through directly to the dynamic config code.
The paste opcode functionality is copied directly from intro.js, with the difference that when the
opcode is entered, we generate a link to start the dynamic config process instead of logging in directly

There is also a new CSS class to ensure that the details/instructions are in a smaller font than the buttons

Testing done:
- Used the "copy and paste" button - popup shows up and we can enter a token
- Scan functionality not tested yet due to lack of support in the emulator; will test on physical phone after all the other functionality is done