have filters show the number of trips in HTML
Instead of a hacky jQuery solution, we can use the 'translate' filter to interpolate in the number of trips shown by the filter, and the total number of trips.
It's still a bit ugly, but at least it's more semantic.
The english labels had &nbsp; (non-breaking-spaces) in them, which were not being parsed correctly. The other languages had all normal spaces, so I made the English spaces all normal too. This text is not at risk of being spread onto two lines.