fix count of displayTrips
After the last commit, untracked time doesn't persist in 'To Label' but it still affects the count X in "Displaying X/Y trips".
Let's just add a counter - we'll track the number of displayed trips in displayTrips.count.
I am hoping we can get rid of the "Displaying X/Y trips" anyway, in preparation for the rewrite-