implement 'logOut()' in general-settings
This used to be `returnToIntro()` - after clicking this, the user could log in with a different OPcode on the same study.
But now that we handle the study/program and codes in one unified token, returning to the intro is pointless.
So let's have this action perform a full logout - it will remove the existing config configuration, including the old token and the old server connection settings.
We'll then trigger a 'hard' refresh (no caching) to reset any cached variables and ensure that views are not going to show old information.
When the app is reloaded, it will start on the 'join' page.

Testing done:
- logout and sign back in with same opcode
- logout and sign back in with different opcode on same study
- logout and sign back in with different opcode on different study
- logout. close app. open app and log in with different opcode