uncommit plugin settings, remove transition-notify
https://github.com/e-mission/e-mission-phone/pull/974#discussion_r1208727550
These files represent settings for various plugins; housed in their respective repos, not here. They should not have been committed. I added them to the .gitignore so that they are not accidentally committed in the future.
I also removed the `transition-notify` files from the script, since it is not being used anymore.