fix + simplify multi-label-ui, scroll resize logic
https://github.com/e-mission/e-mission-phone/pull/974#discussion_r1208402145
Using $broadcast for 'recomputeListEntries' and 'scrollResize' does work. We just needed to include `$rootScope` in the controller and factory definitions.
Now we can remove another use of jQuery because we don't need `getScrollElement()`.
Instead of calling `.trigger('scroll-resize')`, we'll broadcast to `infinite_scroll_list`, where the scrollview will be updated using `$ionicScrollDelegate.resize()`.
This seems like a less hacky way of getting the cards to resize and recompute.