make leaflet-view appear correctly on details page
Prior to this change, the map would grow too large. We want it to only get to 50% of the parent.
The CSS changes here relate to how React Native Web is injecting into the existing DOM structure. It uses a bunch of wrapper divs (of class names starting with 'css-view') and we want to make sure these are getting as big as, but only as big as, the containers they are injecting into.