make createRoot work without duplicating roots
With the `collection-repeat` mechanism of Ionic, the infinite scroll list reuses DOM elements as you scroll. In the case of a large list, this helps to prevent the DOM from getting too big.

However this didn't play nice with our React helpers; we would get the following warning:
"You are calling ReactDOMClient.createRoot() on a container that has already been passed to createRoot() before."

The solution here is to have a Map which keeps track of which DOM elements for which roots have already been created. This way we don't duplicate any roots.
A Map was chosen rather than an object because the key needs to be an object reference (in this case a DOM node).