use react-qr-code to show opcodes
There are countless QR code generation libraries; I chose this one because it works using SVGs rather than <canvas>, which means it will still work when we switch to native rendering and can no longer use HTML.
Because we have to `angularize` components if they are going to be used directly in AngularJS, I created a wrapper around QRCode (www/js/control/QrCode.jsx).
Once the parent components have been converted to React, and the QrCode component no longer needs to be angularized, we can remove the wrapper and just use QRCode directly.

-I also generally cleaned up `saveTokenFile.html` a bit; removed <center> tags and fixed indentation