:bug: Customize the privacy policy when displayed from the profile
Per Google requirements, we display the privacy policy from the control screen
as well, and pass in the `$scope` variable so that it can be configured correctly.
However, most of the customizable text in the policy is from the
`$scope.translated_text` variable, which needs to be filled in with the correct
locale.

And then we added several backward compatibility hacks in
https://github.com/e-mission/e-mission-phone/pull/981

This commit copies over the `translated_text` selection and the backward
compatibility hacks so that the in-app privacy policy is also generated properly