implement .fill-container for embedded RN elements
When react-native-web components are embedded in the DOM we need a bit more control over how they behave. Sometimes we want them to exert their own size on their parent containers, and for the parent container to grow with its contents. Sometimes, we just want the component to fit into its container and not affect the container's size. The former is default; for the latter, we'll use the .fill-container class.

This is what I was doing for the <LeafletView> maps anyway, but I'm now making it more generic so that we can use it in other instances too.