bump enketo-core to v6 from npm; bump leaflet
Surveys were not working anymore on this branch. We have been using a static, old-versioned, hacked-together copy of Enketo checked into manual_lib, and it was not compatible work with the other upgrades that have been made.

Since we are now writing in modern JS (modular structure, with ES imports) I wanted to upgrade to enketo-core@latest (which is being well-maintained and has been modernized).

However, I found that Enketo 7.x has issues when prefilling time inputs.
https://github.com/enketo/enketo-core/issues/979
Since this functionality is essential for us, we will be on 6.x for now. Hopefully we can resolve this on enketo-core and get on enketo-core@latest eventually.

Also, our version of Leaflet was upgraded in the process. We don't actually use any maps in our surveys, but Enketo wants the global "L" as Leaflet 1.x and it was a good idea to upgrade Leaflet anyway.