use local tz to prefill enketo time fields
Enketo doesn't allow time inputs to use any timezone other than the user's current local timezone. It's frustrating for our purposes, but Enketo makes sure whatever the user sees is reflected in their local time.
We previously had added a hack in our copy of the Enketo code to get around this (https://github.com/e-mission/e-mission-phone/commit/0f1051f469cd1bc68e4133e459fb92ef04cdb8e3). It worked ok, but since we're tidying + upgrading dependencies I want to avoid hacking Enketo if possible.
I think the best thing we can do here is just give Enketo what it wants - a timestamp in the current local tz offset - and then sort things back out when we resolve the timestamps upon survey completion.
So `resolveTimestamps()` is just going to ignore the tz offsets in the the survey response and instead use the timezone that it gets from the trip/place.
This works because all we really need is the HH:mm value the user entered in each field. We don't care at all what timezone the user was in when they recorded the activity.