fix survey response ts, if not resolved from response
With the diary and label screen unification, the trip object is no longer going to have `.data.properties`. We just want to get start_ts and end_ts off the trip object directly.
And this could also be a place (if we ever find a use case for 'place confirm' or similar), so we should try enter_ts and exit_ts if start_ts and end_ts aren't there.