use origin_key instead of key in input-matcher
The key of composite trip objects is always going to be `analysis/composite_trip`. The `origin_key` describes whether it's a trip, draft trip, place, or untracked time, and this is what we should check for.
And if it doesn't have an `origin_key` it's not any of those, and should be skipped (this is in the case of the 'header' and 'footer' of the list)