display error if addition does not match any entry
If the timestamps from a survey response were modified such that the response didn't match any trip in view, this would appear.
This would theoretically not happen with proper use of the app, but if there is an issue with our input matching then this would help us be aware of it.
One case where this would happen is if the user enters timestamps that are well before anything in timeline.